Moved the NewIPv4UDPConn function from nclient4/conn_linux.go to server4/conn_linux.go
